+++ src_new/aliases/levi_ataxia/for_levi/levi_062424/zgui_redux/007_(LEVIBARS)_Vital_Bars.lua
@@ -1,0 +1,16 @@
+--[[mudlet
+type: alias
+name: (LEVIBARS) Vital Bars
+hierarchy:
+- Levi_Ataxia
+- Systems
+- zGUI Redux
+attributes:
+  isActive: 'yes'
+  isFolder: 'no'
+regex: ^levibars(?: (.+))?$
+command: ''
+packageName: ''
+]]--
+
+ataxia.bars.dispatch(matches[2])