+++ src_new/triggers/levi_ataxia/for_levi/leviticus/wilderness_map/002_GMCP_Rooms.lua
@@ -33,9 +33,9 @@ mSoundFile: ''
 colorTriggerFgColor: '#000000'
 colorTriggerBgColor: '#000000'
 patterns:
-- pattern: return gmcp.Room.Info.num < 2121028
+- pattern: return gmcp.Room and gmcp.Room.Info and gmcp.Room.Info.num < 2121028
   type: 4
-- pattern: return gmcp.Room.Info.num > 0
+- pattern: return gmcp.Room and gmcp.Room.Info and gmcp.Room.Info.num > 0
   type: 4
 ]]--